packed version detection in own class
