+++ C Sharp Source/LabVIEW CLI/LvLauncher.cs
@@ -53,45 +53,6 @@ namespace LabVIEW_CLI
             return System.Text.RegularExpressions.Regex.IsMatch(launchPath, ".exe$");
         }
 
-        /// <summary>
-        /// Detects installed LabVIEW versions by searching the registry.
-        /// 
-        /// This follows an example from the NI Developer Community (https://decibel.ni.com/content/docs/DOC-25920).
-        /// </summary>
-        /// <returns>a dictionary containing the detected versions and the respective paths</returns>
-        public static Dictionary<string, string> DetectLvVersions()
-        {
-            
-            List<string> excludedKeys = new List<string> { "AddOns", "CurrentVersion" };
-            Dictionary<string, string> versions = new Dictionary<string, string>();
-
-            RegistryKey LvBaseKey = RegistryKey.OpenBaseKey(RegistryHive.LocalMachine, RegistryView.Registry32).OpenSubKey("SOFTWARE\\National Instruments\\LabVIEW");
-            if (LvBaseKey == null)
-            {
-                Console.Error.WriteLine("No installed version of the NI LabVIEW Development System found...");
-                return versions;
-            }
-
-            RegistryKey itemKey;
-            object itemPath;
-            object itemVersion;
-            foreach (var item in LvBaseKey.GetSubKeyNames())
-            {
-                Console.WriteLine("Testing key \"" + item + "\"...");
-                if (excludedKeys.Contains(item))
-                    continue;
-
-                itemKey = LvBaseKey.OpenSubKey(item);
-                itemPath = itemKey.GetValue("Path");
-                itemVersion = itemKey.GetValue("VersionString");
-                if (itemPath != null && itemVersion != null)
-                {
-                    versions.Add(itemVersion.ToString(), itemPath.ToString());
-                    Console.WriteLine("found \"" + itemVersion.ToString() + "\" (" + itemPath.ToString() + ")");
-                }
-            }
-
-            return versions;
-        }
+        
     }
 }
